requirements.txt
requests==2.32.3
python-dotenv==1.0.1
lxml==5.2.2
pytz==2024.1
APScheduler==3.10.4

